FrameCapture: Fix framebuffer bindings in MEC
While capturing an app, I was crashing on the line reading
replayState.getReadFramebuffer()->id()

That's because replayState did not have a read framebuffer
bound, and we dereferenced nullptr. replayState is a default
state that we've created and we're modifying to make it *look*
like apiState.

We should be updating replayState to match the framebuffers
bound in apistate. In this case state*Framebuffer is read from
apiState and is the value we want to propagate.

AFAICT these have been incorrect since they landed in the
initial implementation, but the code is complex:
https://chromium-review.googlesource.com/c/angle/angle/+/1689329

Test: Tower of Fantasy capture
Bug: b/389808881
Change-Id: Ibf7bc56cd385c34139676c5813926f756bb38f42
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6174746
Reviewed-by: Roman Lavrov <romanl@google.com>
Commit-Queue: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Mark Łobodziński <mark@lunarg.com>