Add a ReadPixels test for offset alignment check
An INVALID_OPERATION error is generated if a pixel pack buffer object
is bound and data is not evenly divisible by the number of basic
machine units needed to store in memory the corresponding GL data
type from table 8.4 for the type parameter. Add a test for this.

Bug: angleproject:352963094
Change-Id: I81eda446df5ff29da1326a1f75f5d5aa0825aafd
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6170763
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Shahbaz Youssefi <syoussefi@chromium.org>
Reviewed-by: Amirali Abdolrashidi <abdolrashidi@google.com>