Vulkan: Fix VUID-vkBindBufferMemory-buffer-01444
  Some platforms issue a VVL error regarding memory allocation for an
external buffer. It seems to be due to the fact that the memory is not
allocated as a dedicated memory for the buffer, whereas the buffer
requirements (as reported in vkGetBufferMemoryRequirements2()) show
that it requires a dedicated memory. This causes an issue when binding
the buffer to the memory.

This change explicitly sets the dedicated memory information for the
external buffer, so the memory is regarded as dedicated and be able
to be bound to the buffer without errors.

* Added VkMemoryDedicatedAllocateInfo when calling to allocate memory
  for the external buffer (InitAndroidExternalMemory()).

  * Appended this structure to the extra allocation info via pNext.

* Unsuppressed VUID-vkBindBufferMemory-buffer-01444.

Bug: angleproject:394598758
Change-Id: I0d952cea204e528c23859c74909dbbb0751d0b65
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/6239029
Reviewed-by: Cody Northrop <cnorthrop@google.com>
Reviewed-by: Roman Lavrov <romanl@google.com>
Commit-Queue: Amirali Abdolrashidi <abdolrashidi@google.com>